refactor: Update examples packages
