--- examples/core/src/index.ts
@@ -1,3 +1,0 @@
-export * from './rgbpp';
-export * from './ckb-xudt';
-export * from './utils';
