--- examples/core/src/rgbpp/index.ts
@@ -1,2 +1,0 @@
-export * from './xudt';
-export * from './spore';
