+++ examples/rgbpp/spore/launch/3-create-spores.ts
@@ -20,7 +20,7 @@ import {
   btcService,
   CKB_PRIVATE_KEY,
   ckbAddress,
-} from 'examples-core';
+} from '../../core';
 
 interface Params {
   clusterRgbppLockArgs: Hex;
