+++ examples/rgbpp/xudt/launch/1-prepare-launch.ts
@@ -13,7 +13,7 @@ import {
   getSecp256k1CellDep,
 } from '@rgbpp-sdk/ckb';
 import { RGBPP_TOKEN_INFO } from './0-rgbpp-token-info';
-import { CKB_PRIVATE_KEY, ckbAddress, collector, isMainnet } from 'examples-core';
+import { CKB_PRIVATE_KEY, ckbAddress, collector, isMainnet } from '../../core';
 
 const prepareLaunchCell = async ({
   outIndex,
