+++ examples/xudt/2-transfer-xudt.ts
@@ -1,5 +1,5 @@
 import { RgbppTokenInfo } from '@rgbpp-sdk/ckb';
-import { transferXudt } from 'examples-core';
+import { transferXudt } from './core/xudt';
 
 const XUDT_TOKEN_INFO: RgbppTokenInfo = {
   decimal: 8,
