+++ examples/xudt/core/env.ts
@@ -1,0 +1,18 @@
+import { AddressPrefix, privateKeyToAddress } from '@nervosnetwork/ckb-sdk-utils';
+import { Collector } from '@rgbpp-sdk/ckb';
+import dotenv from 'dotenv';
+
+dotenv.config({ path: __dirname + '/../.env' });
+
+export const isMainnet = process.env.IS_MAINNET === 'true' ? true : false;
+
+console.log('IS_MAINNET', process.env.IS_MAINNET);
+
+export const collector = new Collector({
+  ckbNodeUrl: process.env.CKB_NODE_URL!,
+  ckbIndexerUrl: process.env.CKB_INDEXER_URL!,
+});
+export const CKB_PRIVATE_KEY = process.env.CKB_SECP256K1_PRIVATE_KEY!;
+export const ckbAddress = privateKeyToAddress(CKB_PRIVATE_KEY, {
+  prefix: isMainnet ? AddressPrefix.Mainnet : AddressPrefix.Testnet,
+});
