+++ examples/xudt/package.json
@@ -5,12 +5,17 @@
   "private": true,
   "type": "commonjs",
   "scripts": {
+    "build": "tsc -p tsconfig.build.json",
     "format": "prettier --write '**/*.ts'",
     "lint": "tsc && eslint . && prettier --check '**/*.ts'",
     "lint:fix": "tsc && eslint --fix --ext .ts . && prettier --write '**/*.ts'"
   },
   "dependencies": {
-    "@rgbpp-sdk/ckb": "workspace:^",
-    "examples-core": "workspace:^"
+    "@nervosnetwork/ckb-sdk-utils": "^0.109.1",
+    "@rgbpp-sdk/ckb": "workspace:^"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.5",
+    "@types/dotenv": "^8.2.0"
   }
 }
