+++ examples/xudt/tsconfig.build.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "CommonJS",
+    "rootDir": "src",
+    "outDir": "lib",
+    "noEmit": false
+  },
+  "include": ["core/*.ts"],
+}
